COMMERCE-9599 Moved product options js implementation into another module
