+++ modules/apps/commerce/commerce-frontend-js/src/main/resources/META-INF/resources/components/product_options/utils.js
